refactor(ui): abstract components HBL-174
