+++ libs/ui/components/src/lib/Anchor/Anchor.spec.tsx
@@ -1,0 +1,10 @@
+import { render } from '@testing-library/react';
+
+import Anchor from './Anchor';
+
+describe('<Anchor />', () => {
+  it('should render properly', () => {
+    const { container } = render(<Anchor />);
+    expect(container).toBeInTheDocument();
+  });
+});
