+++ libs/ui/components/src/lib/Anchor/Anchor.tsx
@@ -1,0 +1,3 @@
+import { styled } from "@mui/material";
+
+export default styled('a')({ textDecoration: 'none', backroundColor: 'none' });
