+++ libs/ui/components/src/lib/Anchor/index.ts
@@ -1,0 +1,3 @@
+import Anchor from './Anchor';
+
+export default Anchor;
