+++ libs/ui/components/src/lib/Dashboard/DashboardGymZone/DashboardGymZone.spec.tsx
@@ -21,50 +21,6 @@ describe('<DashboardGymZone />', () => {
     expect(container).toBeInTheDocument();
   });
 
-  describe('gymZone', () => {
-    it('should render the gym fields', () => {
-      render(<DashboardGymZone gymZone={gymZone as any} />);
-
-      expect(screen.getByText(gymZone.name.toUpperCase())).toBeInTheDocument();
-      expect(screen.getByText(gymZone.description)).toBeInTheDocument();
-      expect(
-        screen.getByText(`${gymZone.openTime} - ${gymZone.closeTime}`)
-      ).toBeInTheDocument();
-      expect(
-        screen.getByText(`Maximum capacity: ${gymZone.capacity}`)
-      ).toBeInTheDocument();
-
-      // Find icons
-      expect(screen.getByLabelText('Non class zone')).toBeInTheDocument();
-      expect(screen.getByTitle('non-class-zone')).toBeInTheDocument();
-      expect(screen.getByLabelText('Facial mask required')).toBeInTheDocument();
-      expect(screen.getByTitle('mask-required')).toBeInTheDocument();
-      expect(
-        screen.getByLabelText('Covid passport required')
-      ).toBeInTheDocument();
-      expect(screen.getByTitle('passport-required')).toBeInTheDocument();
-    });
-
-    it('should render the opposite icon titles', () => {
-      render(
-        <DashboardGymZone
-          gymZone={
-            {
-              ...gymZone,
-              isClassType: true,
-              covidPassport: false,
-              maskRequired: false
-            } as any
-          }
-        />
-      );
-
-      expect(screen.getByTitle('class-zone')).toBeInTheDocument();
-      expect(screen.getByTitle('mask-not-required')).toBeInTheDocument();
-      expect(screen.getByTitle('passport-not-required')).toBeInTheDocument();
-    });
-  });
-
   describe('onClick', () => {
     it('should call onClick if passed', () => {
       const onClickSpy = jest.fn();
