+++ libs/ui/components/src/lib/Dashboard/DashboardGymZone/DashboardGymZone.tsx
@@ -1,30 +1,14 @@
 import React from 'react';
 
 import { GymZoneDTO } from '@hubbl/shared/models/dto';
-import { CallToAction, FitnessCenter, Masks } from '@mui/icons-material';
-import {
-  CardActionArea,
-  Stack,
-  styled,
-  Tooltip,
-  Typography
-} from '@mui/material';
+import { CardActionArea, styled } from '@mui/material';
 
-import ContentCard from '../../ContentCard';
+import Anchor from '../../Anchor';
+import GymZoneListItem from '../../GymZoneListItem';
 
 const ContentCardAction = styled(CardActionArea)(({ theme }) => ({
   borderRadius: theme.spacing(2)
 }));
-
-const PaddedContentCard = styled(ContentCard)(({ theme }) => ({
-  padding: theme.spacing(2, 3),
-  width: theme.spacing(44),
-  height: theme.spacing(25)
-}));
-
-const ContentStack = styled(Stack)({ height: '100%' });
-
-const Anchor = styled('a')({ textDecoration: 'none', backroundColor: 'none' });
 
 /**
  * Props that are required to be passed to the `a` component,
@@ -49,77 +33,7 @@ const DashboardGymZone = React.forwardRef<
   ({ gymZone, href, onClick }, ref): JSX.Element => (
     <Anchor href={href} ref={ref} onClick={onClick}>
       <ContentCardAction color="primary">
-        <PaddedContentCard>
-          <ContentStack justifyContent="space-between">
-            <Stack gap={1}>
-              <Stack
-                direction="row"
-                justifyContent="space-between"
-                alignItems="center"
-              >
-                <Typography variant="h6">
-                  {gymZone.name.toUpperCase()}
-                </Typography>
-
-                <Stack gap={1} direction="row" alignItems="center">
-                  <Tooltip
-                    title={
-                      gymZone.isClassType ? 'Class zone' : 'Non class zone'
-                    }
-                  >
-                    <FitnessCenter
-                      sx={{ transform: 'rotate(-45deg)' }}
-                      color={gymZone.isClassType ? 'success' : undefined}
-                      titleAccess={
-                        gymZone.isClassType ? 'class-zone' : 'non-class-zone'
-                      }
-                    />
-                  </Tooltip>
-
-                  <Tooltip
-                    title={`Facial mask${
-                      gymZone.maskRequired ? '' : 'not'
-                    } required`}
-                  >
-                    <Masks
-                      sx={{ fontSize: '1.75rem' }}
-                      color={gymZone.maskRequired ? 'success' : undefined}
-                      titleAccess={
-                        gymZone.maskRequired
-                          ? 'mask-required'
-                          : 'mask-not-required'
-                      }
-                    />
-                  </Tooltip>
-
-                  <Tooltip
-                    title={`Covid passport${
-                      gymZone.covidPassport ? '' : 'not'
-                    } required`}
-                  >
-                    <CallToAction
-                      color={gymZone.covidPassport ? 'success' : undefined}
-                      titleAccess={
-                        gymZone.covidPassport
-                          ? 'passport-required'
-                          : 'passport-not-required'
-                      }
-                    />
-                  </Tooltip>
-                </Stack>
-              </Stack>
-
-              <Typography>{gymZone.description}</Typography>
-            </Stack>
-
-            <Stack gap={1}>
-              <Typography>
-                {gymZone.openTime} - {gymZone.closeTime}
-              </Typography>
-              <Typography>Maximum capacity: {gymZone.capacity}</Typography>
-            </Stack>
-          </ContentStack>
-        </PaddedContentCard>
+        <GymZoneListItem gymZone={gymZone} />
       </ContentCardAction>
     </Anchor>
   )
