+++ libs/ui/components/src/lib/GymZoneListItem/index.ts
@@ -1,0 +1,3 @@
+import GymZoneListItem from './GymZoneListItem';
+
+export default GymZoneListItem;
